Player.cs
Player.cs에 직업 추가
직업 enum 추가
직업마다 스탯 변화 추가
치명타 시스템 수정
회피 시스템 수정

+++ TeamSpartaDungeonGame/TeamSpartaDungeonGame/Interface/IAction.cs
@@ -12,7 +12,7 @@ namespace TeamSpartaDungeonGame.Interface
         public void Attack();
         public void Death();
         public void UsingItem();
-        public int Critical();
+        public float Critical();
         public void Dodge();
     }
 }
